fix flag data size for ready status display ready status between holes on score board

+++ samples/golf/src/golf/BilliardsState.cpp
@@ -1301,7 +1301,7 @@ void BilliardsState::handleNetEvent(const net::NetEvent& evt)
             showReadyNotify(evt.packet.as<BilliardsPlayer>());
             break;
         case PacketID::ReadyQuitStatus:
-            m_readyQuitFlags = evt.packet.as<std::uint8_t>();
+            m_readyQuitFlags = evt.packet.as<std::uint16_t>();
             break;
         case PacketID::GameEnd:
             showGameEnd(evt.packet.as<BilliardsPlayer>());
